worker: Improve pasting behavior
Pasting a source file or folder that already exists in
the  target  directory will result in a renamed source
file or folder.

e.g. If copying or moving a  "New Folder"  folder into a
     directory where another "New Folder" already exists,
     it will be pasted with  "New Folder(1)" as its name.

This also fixes the issue where moving a directory would
show up a file, the first time.