worker: Show message on folder load error
In some cases the OS report directories that doesn't exists,
or the user don't have the required permissions to read them,
in that case the call to 'os.listdir()' will fail with an
OSError.

This commits add a new event to the PorfolioLoadWorker that
is fired if the directory can't be listed.

This new event is handled and a message with a summary
is displayed to the user.